Move Java metadata classes into their own top level namespace (outside .common)
