+++ trac-api/trac-metadata/src/main/proto/trac/metadata/data.proto
@@ -17,7 +17,7 @@
 syntax = 'proto3';
 package trac.metadata;
 
-option java_package = "com.accenture.trac.common.metadata";
+option java_package = "com.accenture.trac.metadata";
 option java_multiple_files = true;
 
 import "trac/metadata/type.proto";
