+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/api/BaseMetadataApi.java
@@ -17,7 +17,7 @@
 package com.accenture.trac.svc.meta.api;
 
 import com.accenture.trac.common.api.*;
-import com.accenture.trac.common.metadata.*;
+import com.accenture.trac.metadata.*;
 import com.accenture.trac.common.util.ApiWrapper;
 
 import com.accenture.trac.svc.meta.services.MetadataReadService;
