+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/services/MetadataReadService.java
@@ -16,10 +16,10 @@
 
 package com.accenture.trac.svc.meta.services;
 
-import com.accenture.trac.common.metadata.TagSelector;
+import com.accenture.trac.metadata.TagSelector;
+import com.accenture.trac.metadata.ObjectType;
+import com.accenture.trac.metadata.Tag;
 import com.accenture.trac.svc.meta.dal.IMetadataDal;
-import com.accenture.trac.common.metadata.ObjectType;
-import com.accenture.trac.common.metadata.Tag;
 
 import java.util.List;
 import java.util.UUID;
