+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/services/MetadataSearchService.java
@@ -16,8 +16,8 @@
 
 package com.accenture.trac.svc.meta.services;
 
-import com.accenture.trac.common.metadata.Tag;
-import com.accenture.trac.common.metadata.search.SearchParameters;
+import com.accenture.trac.metadata.Tag;
+import com.accenture.trac.metadata.search.SearchParameters;
 import com.accenture.trac.svc.meta.dal.IMetadataDal;
 
 import java.util.List;
