+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/services/MetadataWriteService.java
@@ -16,7 +16,8 @@
 
 package com.accenture.trac.svc.meta.services;
 
-import com.accenture.trac.common.metadata.*;
+import com.accenture.trac.metadata.*;
+import com.accenture.trac.common.metadata.MetadataCodec;
 import com.accenture.trac.svc.meta.dal.IMetadataDal;
 import com.accenture.trac.svc.meta.validation.MetadataValidator;
