+++ trac-services/trac-svc-meta/src/test/java/com/accenture/trac/svc/meta/dal/MetadataDalEncodingTest.java
@@ -16,9 +16,9 @@
 
 package com.accenture.trac.svc.meta.dal;
 
-import com.accenture.trac.common.metadata.BasicType;
+import com.accenture.trac.metadata.BasicType;
+import com.accenture.trac.metadata.ObjectType;
 import com.accenture.trac.common.metadata.MetadataCodec;
-import com.accenture.trac.common.metadata.ObjectType;
 import static com.accenture.trac.svc.meta.test.TestData.*;
 
 import java.math.BigDecimal;
