+++ trac-services/trac-svc-meta/src/test/java/com/accenture/trac/svc/meta/services/TagUpdateServiceTest.java
@@ -16,7 +16,9 @@
 
 package com.accenture.trac.svc.meta.services;
 
-import com.accenture.trac.common.metadata.*;
+import com.accenture.trac.metadata.*;
+import com.accenture.trac.common.metadata.TypeSystem;
+import com.accenture.trac.common.metadata.MetadataCodec;
 import com.accenture.trac.svc.meta.exception.ETagUpdate;
 
 import org.junit.jupiter.api.Test;
